Minor type improvements to example
The example now uses some simple type-hinting for the TopCodes array,
and the result is no longer parsed as JSON (since this was not
necessary). I'm avoiding TypeScript in the example as it's quite small
at the moment, but it may be worth migrating over to use the example
as an integration test of all the exports from the WASM build.